chore: split code
